recommender-ai-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
requests>=2.31
django-cors-headers>=4.3
psycopg2-binary>=2.9
scikit-learn>=1.3
numpy>=1.24
